lcd: change eDP pixel clk calculate strategy [1/1]
PD#SWPL-139153

Problem:
current eDP pclk calculate method may fail due to min error too small

Solution:
change pclk calculate strategy

Verify:
t7_an400

+++ include/amlogic/media/vout/lcd/lcd_vout.h
@@ -150,7 +150,7 @@ struct lcd_timing_s {
 				* [7:0]: ss_level
 				*/
 
-	unsigned short sync_duration_num;
+	unsigned int sync_duration_num;
 	unsigned short sync_duration_den;
 
 	unsigned int hstart;
